refactor: split generate.ts into cli/ and actions/ modules (#113)
