+++ src/actions/plan.ts
@@ -1,0 +1,6 @@
+import { formatOnboardingPlan } from '../planner.js';
+import type { OnboardingPlan } from '../planner.js';
+
+export function runPlanAction(onboardingPlan: OnboardingPlan): void {
+  console.log(formatOnboardingPlan(onboardingPlan));
+}
